hpke::danger

Module streaming_enc

The underlying streaming/online encryption primitive defined in the HPKE spec.

DO NOT use the same key for two sender contexts or two receiver contexts. Doing this can lead to reflection attacks, i.e., replaying Alice’s message to Alice herself, pretending it came from Bob.

Example use:

// These types define the ciphersuite Alice and Bob will be using
type Kem = XWing;
type Aead = ChaCha20Poly1305;
type Kdf = KdfTurboShake128;

let mut csprng = rand::rng();

// This is a description string for the session. Both Alice and Bob need to know this value.
// It's not secret.
let info_str = b"Alice and Bob's weekly chat";

// Pick a random key, base nonce, and exporter secret for the streaming encryption context
let mut key = AeadKey::<Aead>::default();
let mut base_nonce = AeadNonce::default();
let mut exporter_secret = ExporterSecret::<Kdf>::default();
csprng.fill_bytes(key.0.as_mut_slice());
csprng.fill_bytes(base_nonce.0.as_mut_slice());
csprng.fill_bytes(exporter_secret.0.as_mut_slice());

// Set up the sender and receiver contexts
let base_nonce_copy = AeadNonce::<Aead>(base_nonce.0);
let exporter_secret_copy = ExporterSecret::<Kdf>(exporter_secret.0);
let mut sender_ctx = create_sender_context::<_, _, Kem>(
    &key,
    base_nonce_copy,
    exporter_secret_copy,
);
let mut receiver_ctx = create_receiver_context::<_, _, Kem>(
    &key,
    base_nonce,
    exporter_secret,
);

let msg = b"fronthand or backhand?";
let aad = b"a gentleman's game";

// Encrypt in place with the sender context
let ciphertext = sender_ctx.seal(msg, aad).expect("seal() failed");
// Decrypt with the receiver context
let decrypted = receiver_ctx.open(&ciphertext, aad).expect("open() failed");

// Check the decrypted message was what was sent
assert_eq!(&decrypted, msg);

Structs

AeadKey

A struct representing a generic key for an AEAD cipher. Implements Default and Zeroize, and zeroizes on drop.

AeadNonce

A nonce is a bytestring you only use for encryption once. Implements Default and Zeroize, and zeroizes on drop.

ExporterSecret

Secret generated in derive_enc_ctx and stored in AeadCtx. Implements Default and Zeroize, and zeroizes on drop.

Functions

create_receiver_context

Creates a streaming encryption receiver context from a key, nonce, and exporter secret.

create_sender_context

Creates a streaming encryption sender context from a key, nonce, and exporter secret.