hpke::danger::streaming_enc

Function create_sender_context

pub fn create_sender_context<A: Aead, Kdf: KdfTrait, Kem: KemTrait>(
    key: &AeadKey<A>,
    base_nonce: AeadNonce<A>,
    exporter_secret: ExporterSecret<Kdf>,
) -> AeadCtxS<A, Kdf, Kem>

Creates a streaming encryption sender context from a key, nonce, and exporter secret.

⚠️ DANGER

This is a low-level API. Only use this if you know what you are doing.

Use this method to set up an online/streaming encryption channel with shared secrets derived via some key exchange. This is what the HPKE spec does when it instantiates a ContextS using (a hash of) the KEM shared secret.

This method can also be used to create a response context like described in section 9.8 of the HPKE spec.