hpke

Function setup_sender_with_rng

pub fn setup_sender_with_rng<A, Kdf, Kem>(
    mode: &OpModeS<'_, Kem>,
    pk_recip: &Kem::PublicKey,
    info: &[u8],
    csprng: &mut impl CryptoRng,
) -> Result<(Kem::EncappedKey, AeadCtxS<A, Kdf, Kem>), HpkeError>
where
    A: Aead,
    Kdf: KdfTrait,
    Kem: KemTrait,

Initiates an encryption context to the given recipient public key. info is a domain separator.

NOTE: The post-quantum KEMs (XWing, MlKem768P256, MlKem1024P384, MlKem768, and MlKem1024) do not support authenticated encapsulation, so mode MUST be Base or Psk.

Return Value

On success, returns an encapsulated public key (intended to be sent to the recipient), and an encryption context. If an error happened during key encapsulation, returns Err(HpkeError::EncapError).

Panics

Panics if mode is not Base or Psk, or if info.len() + mode.get_psk_id().len() + 5 ≥ 2¹⁶.