hpke

Function single_shot_seal_inout_detached

pub fn single_shot_seal_inout_detached<A, Kdf, Kem>(
    mode: &OpModeS<'_, Kem>,
    pk_recip: &Kem::PublicKey,
    info: &[u8],
    buffer: InOutBuf<'_, '_, u8>,
    aad: &[u8],
) -> Result<(Kem::EncappedKey, AeadTag<A>), HpkeError>
where
    A: Aead,
    Kdf: KdfTrait,
    Kem: KemTrait,

Does a crate::setup_sender and AeadCtxS::seal_inout_detached in one shot. That is, it does a key encapsulation to the specified recipient and encrypts the provided plaintext in place.

Return Value

Returns Ok((encapped_key, auth_tag)) on success. If an error happened during key encapsulation, returns Err(HpkeError::EncapError). If an error happened during encryption, returns Err(HpkeError::SealError). In this case, the contents of plaintext is undefined.

Panics

Panics if getrandom::SysRng fails to generate random bytes.