hpke

Trait Kem

pub trait Kem: Sized {
    type PublicKey: Clone + Debug + PartialEq + Eq + Serializable + Deserializable;
    type PrivateKey: Clone + ConstantTimeEq + Serializable + Deserializable;
    type EncappedKey: Clone + Serializable + Deserializable;

    const KEM_ID: u16;

    // Required methods
    fn sk_to_pk(sk: &Self::PrivateKey) -> Self::PublicKey;
    fn derive_keypair(ikm: &[u8]) -> (Self::PrivateKey, Self::PublicKey);
    fn encap_with_rng(
        pk_recip: &Self::PublicKey,
        sender_id_keypair: Option<(&Self::PrivateKey, &Self::PublicKey)>,
        csprng: &mut impl CryptoRng,
    ) -> Result<(SharedSecret<Self>, Self::EncappedKey), HpkeError>;

    // Provided methods
    fn gen_keypair() -> (Self::PrivateKey, Self::PublicKey) { ... }
    fn gen_keypair_with_rng(
        csprng: &mut impl CryptoRng,
    ) -> (Self::PrivateKey, Self::PublicKey) { ... }
    fn encap(
        pk_recip: &Self::PublicKey,
        sender_id_keypair: Option<(&Self::PrivateKey, &Self::PublicKey)>,
    ) -> Result<(SharedSecret<Self>, Self::EncappedKey), HpkeError> { ... }
}

Represents authenticated encryption functionality

Required Associated Constants

const KEM_ID: u16

The algorithm identifier for a KEM implementation

Required Associated Types

type PublicKey: Clone + Debug + PartialEq + Eq + Serializable + Deserializable

The key exchange’s public key type. If you want to generate a keypair, see Kem::gen_keypair or Kem::derive_keypair

type PrivateKey: Clone + ConstantTimeEq + Serializable + Deserializable

The key exchange’s private key type. If you want to generate a keypair, see Kem::gen_keypair or Kem::derive_keypair

type EncappedKey: Clone + Serializable + Deserializable

The encapsulated key for this KEM. This is used by the recipient to derive the shared secret.

Required Methods

fn sk_to_pk(sk: &Self::PrivateKey) -> Self::PublicKey

Computes the public key of a given private key

fn derive_keypair(ikm: &[u8]) -> (Self::PrivateKey, Self::PublicKey)

Deterministically derives a keypair from the given input keying material

Requirements

This keying material SHOULD have as many bits of entropy as the bit length of a secret key, i.e., 8 * Self::PrivateKey::size(). For X25519 and P-256, this is 256 bits of entropy.

fn encap_with_rng( pk_recip: &Self::PublicKey, sender_id_keypair: Option<(&Self::PrivateKey, &Self::PublicKey)>, csprng: &mut impl CryptoRng, ) -> Result<(SharedSecret<Self>, Self::EncappedKey), HpkeError>

Derives a shared secret and an ephemeral pubkey that the owner of the reciepint’s pubkey can use to derive the same shared secret. If sk_sender_id is given, the sender’s identity will be tied to the shared secret. All this does is generate an ephemeral keypair and pass to encap_with_eph.

Return Value

Returns a shared secret and encapped key on success. If an error happened during key exchange, returns Err(HpkeError::EncapError).

Provided Methods

fn gen_keypair() -> (Self::PrivateKey, Self::PublicKey)

Generates a random keypair using the system RNG.

Panics

Panics if getrandom::SysRng fails to generate random bytes.

fn gen_keypair_with_rng( csprng: &mut impl CryptoRng, ) -> (Self::PrivateKey, Self::PublicKey)

Generates a random keypair using the given RNG

fn encap( pk_recip: &Self::PublicKey, sender_id_keypair: Option<(&Self::PrivateKey, &Self::PublicKey)>, ) -> Result<(SharedSecret<Self>, Self::EncappedKey), HpkeError>

Derives a shared secret and an ephemeral pubkey that the owner of the reciepint’s pubkey can use to derive the same shared secret. If sk_sender_id is given, the sender’s identity will be tied to the shared secret. All this does is generate an ephemeral keypair and pass to encap_with_eph.

Return Value

Returns a shared secret and encapped key on success. If an error happened during key exchange, returns Err(HpkeError::EncapError).

Panics

Panics if getrandom::SysRng fails to generate random bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.